#6 switched to ninit framework
