+++ Phaethon/Tests/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="EntityFramework" version="6.2.0" targetFramework="net461" />
-  <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net461" />
-  <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net461" />
+  <package id="NUnit" version="3.11.0" targetFramework="net461" />
 </packages>
